PMM-544 fixes & update grafana packages
