+++ pmm-app/package.json
@@ -31,9 +31,9 @@
     "grafana"
   ],
   "dependencies": {
-    "@grafana/data": "7.0.0",
-    "@grafana/runtime": "7.0.0",
-    "@grafana/ui": "7.0.0",
+    "@grafana/data": "7.0.3",
+    "@grafana/runtime": "7.0.3",
+    "@grafana/ui": "7.0.3",
     "antd": "^3.24.3",
     "axios": "^0.19.0",
     "emotion": "10.0.27",
@@ -49,7 +49,7 @@
   },
   "devDependencies": {
     "@babel/preset-env": "^7.9.5",
-    "@grafana/toolkit": "7.0.0",
+    "@grafana/toolkit": "7.0.3",
     "@types/enzyme": "^3.10.5",
     "@types/enzyme-adapter-react-16": "^1.0.6",
     "@types/grafana": "github:CorpGlory/types-grafana.git",
